Move VotingBox files to Items directory
