+++ client/src/modules/list/components/Items/ItemsList.jsx
@@ -5,7 +5,7 @@ import { withRouter } from 'react-router-dom';
 
 import ListItem from 'modules/list/components/Items/ListItem';
 import { getCurrentUser } from 'modules/authorization/model/selectors';
-import { toggle, clearVote, setVote } from './actions';
+import { toggle, clearVote, setVote } from './model/actions';
 import { RouterMatchPropType, UserPropType } from 'common/constants/propTypes';
 
 const DISPLAY_LIMIT = 3;
