+++ client/src/modules/list/components/Items/ListItem/index.js
@@ -1,0 +1,3 @@
+import ListItem from './ListItem';
+
+export default ListItem;
