+++ client/src/modules/list/components/Items/VotingBox/VotingBox.scss
